db_requirements.txt
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
passlib>=1.7.4
python-dotenv>=0.19.0
alembic>=1.7.0
bcrypt>=3.2.0
